# openapi.yaml
openapi: 3.0.3
info: { title: NovelGrain Backend API, version: 0.4.0 }
servers: [ { url: /api } ]
paths:
  /auth/phone/login:
    post: { summary: 手机登录, responses: { '200': { description: OK } } }
  /auth/wechat/login:
    post: { summary: 微信登录, responses: { '200': { description: OK } } }
  /me:
    get: { summary: 当前用户, responses: { '200': { description: OK } } }
    patch: { summary: 更新当前用户, responses: { '200': { description: OK } } }
  /books:
    get:
      summary: 列表
      parameters:
        - { in: query, name: tab, schema: { type: string, enum: [ hot, new ] } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: orientation, schema: { type: string } }
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: tag, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: size, schema: { type: integer, default: 20 } }
      responses: { '200': { description: OK } }
    post:
      summary: 新增
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BookCreate' } } } }
      responses: { '200': { description: OK } }
  /books/{id}:
    get: { summary: 详情, parameters: [ { in: path, name: id, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
    patch:
      summary: 更新
      parameters: [ { in: path, name: id, required: true, schema: { type: integer } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BookUpdate' } } } }
      responses: { '200': { description: OK } }
    delete:
      summary: 删除
      parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ]
      responses: { '200': { description: OK } }
  /books/{id}/likes:
    post: { summary: 点赞, parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
    delete: { summary: 取消点赞, parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
  /books/{id}/bookmarks:
    post: { summary: 收藏, parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
    delete: { summary: 取消收藏, parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
  /books/{id}/comments:
    get:
      summary: 评论列表
      parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: page, schema: { type: integer, default: 1 } }, { in: query, name: size, schema: { type: integer, default: 20 } } ]
      responses: { '200': { description: OK } }
    post:
      summary: 新增评论
      parameters: [ { in: path, name: id, required: true, schema: { type: integer } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CommentCreate' } } } }
      responses: { '200': { description: OK } }
  /tags/suggest:
    get: { summary: 标签建议, parameters: [ { in: query, name: q, required: true, schema: { type: string } } ], responses: { '200': { description: OK } } }
  /tags:
    post: { summary: 创建标签, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TagCreate' } } } }, responses: { '200': { description: OK } } }
  /leaderboard:
    get: { summary: 排行榜, parameters: [ { in: query, name: type, schema: { type: string, enum: [ champion, rookie ] } }, { in: query, name: limit, schema: { type: integer, default: 10 } } ], responses: { '200': { description: OK } } }
  /notifications:
    get: { summary: 通知列表, parameters: [ { in: query, name: userId, required: true, schema: { type: integer } }, { in: query, name: page, schema: { type: integer, default: 1 } }, { in: query, name: size, schema: { type: integer, default: 20 } } ], responses: { '200': { description: OK } } }
  /notifications/read-all:
    post: { summary: 全部已读, parameters: [ { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }
  /notifications/{id}/read:
    post: { summary: 单条已读, parameters: [ { in: path, name: id, required: true, schema: { type: integer } }, { in: query, name: userId, required: true, schema: { type: integer } } ], responses: { '200': { description: OK } } }

components:
  schemas:
    BookCreate:
      type: object
      required: [ title, orientation, category ]
      properties:
        title: { type: string }
        author: { type: string }
        orientation: { type: string }
        category: { type: string }
        blurb: { type: string }
        summary: { type: string }
        tags: { type: array, items: { type: string } }
        coverUrl: { type: string }
        recommenderId: { type: integer }
    BookUpdate:
      type: object
      properties:
        title: { type: string }
        author: { type: string }
        orientation: { type: string }
        category: { type: string }
        blurb: { type: string }
        summary: { type: string }
        coverUrl: { type: string }
        tags: { type: array, items: { type: string } }
    CommentCreate:
      type: object
      required: [ text, userId ]
      properties:
        text: { type: string }
        parentId: { type: integer, nullable: true }
        userId: { type: integer }
    TagCreate:
      type: object
      required: [ name ]
      properties: { name: { type: string } }